refactor: removed services folder for standardization
