+++ index.html
@@ -1,10 +1,18 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <title>NUSTUTS</title>
+    <meta
+      name="description"
+      content="Welcome to NUSTUTS! A centralized platform tailored for Teaching Assistants (TAs) to share
+                materials, files and links with Students in their tutorials. Seamlessly
+                book consultations with TAs, track attendance and have meaningful
+                discussions with one another. Maximise your learning experience by
+                joining now!"
+    />
   </head>
   <body>
     <div id="root"></div>
