+++ src/components/consultations/Consultations.tsx
@@ -3,8 +3,9 @@ import { Consultation } from "../../types";
 import { useAuthContext } from "../../hooks/useAuthContext";
 import ConsultationBooking from "./ConsultationBooking";
 import { isCurrentDateTimePastGivenDateTime } from "../../util/util";
-import { getConsultationsForDate } from "../../services/consultations";
 import LoadingSpinner from "../LoadingSpinner";
+import { AuthenticatedUser } from "../../context/AuthContext";
+import axios from "axios";
 
 type ConsultationBookingClickable = {
   consultationData: Consultation;
@@ -23,6 +24,7 @@ const Consultations = ({
     ConsultationBookingClickable[]
   >([]);
   const [isLoading, setIsLoading] = useState(true);
+  const BASE_URL = import.meta.env.VITE_BASE_URL;
 
   useEffect(() => {
     const fetchConsultations = async () => {
@@ -57,20 +59,40 @@ const Consultations = ({
     fetchConsultations();
   }, [tutorialId]);
 
+  const getConsultationsForDate = async (
+    tutorialId: number,
+    date: string,
+    user: AuthenticatedUser
+  ): Promise<Consultation[]> => {
+    try {
+      const res = await axios.get(
+        `${BASE_URL}/api/consultations/${tutorialId}`,
+        {
+          params: { date: date },
+          headers: { Authorization: `Bearer ${user.tokens.accessToken}` },
+        }
+      );
+      return res.data.data.consultations;
+    } catch (error) {
+      console.log(error);
+      return [];
+    }
+  };
+
   return isLoading ? (
     <LoadingSpinner />
   ) : consultations.filter((consultation) => consultation.isClickable).length >
     0 ? (
-  <div className="p-4">
-    <h1 className="text-2xl font-semibold mb-4 border-b pb-2">{`Consultations for ${date}`}</h1>
-    {consultations.map((consultation) => (
-      <ConsultationBooking
-        key={consultation.consultationData.id}
-        consultationData={consultation.consultationData}
-        isClickable={consultation.isClickable}
-      />
-    ))}
-  </div>
+    <div className="p-4">
+      <h1 className="text-2xl font-semibold mb-4 border-b pb-2">{`Consultations for ${date}`}</h1>
+      {consultations.map((consultation) => (
+        <ConsultationBooking
+          key={consultation.consultationData.id}
+          consultationData={consultation.consultationData}
+          isClickable={consultation.isClickable}
+        />
+      ))}
+    </div>
   ) : (
     <div className="p-4">
       <h1 className="text-2xl font-semibold mb-4 border-b pb-2">{`Consultations for ${date}`}</h1>
